DO NOT MERGE: F2FS: swithc /data and /cache to F2FS
FOR TESTING PURPOSES ONLY:
This is just an example to convert data and cache to f2fs file system.
This will never be merged because it requires a full data wipe.

This is however does work with matching kernel commits to enable F2FS
and a compatable recovery built with f2fs-tools and kernel support.

Change-Id: I0c14df7af278507b9fb85260497a8eef8991faf4